+++ cocos2dx/textures/CCTextureCache.h
@@ -140,7 +140,7 @@ public:
      * The "filenName" parameter is the related/absolute path of the file image.
      * Return true and image width and Height if the reloading is succeed, otherwise return false.
      */
-    bool reloadTexture(const char* fileName, unsigned int &imageWidth, unsigned int &imageHeight);
+    bool reloadTexture(const char* fileName, unsigned short &imageWidth, unsigned short &imageHeight);
 
     /** Purges the dictionary of loaded textures.
     * Call this method if you receive the "Memory Warning"
